+++ src/Klotski.utility.ts
@@ -1,0 +1,13 @@
+const shuffleTiles = (size: number) =>
+  Array(size ** 2)
+    .fill(null)
+    .map((_, index) => index)
+    .reduce(
+      (carry: number[][], _, index, original) =>
+        index % size === 0
+          ? [...carry, ...[original.slice(index, index + size)]]
+          : carry,
+      []
+    );
+
+export { shuffleTiles };
